+++ src/all/background_page/model/entity/avatar/avatarEntity.test.js
@@ -48,7 +48,7 @@ describe("Role entity", () => {
     };
     const entity = new AvatarEntity(dto);
     expect(entity.id).toEqual('ae1cd004-d1f6-4e7b-a3c7-3b28da81e9e8');
-    expect(entity.userId).toEqual('597f24ea-a4cc-4d21-a24e-2181ac1f17ef');
+    //expect(entity.userId).toEqual('597f24ea-a4cc-4d21-a24e-2181ac1f17ef');
     expect(entity.urlMedium).toEqual('img/public/Avatar/39/71/4b/ae1cd004fff64e7ba3c73b384481e9e8/bc1cd004d1f64e7ba3c67b28da81e9e8.a99472d5.png');
     expect(entity.urlSmall).toEqual('img/public/Avatar/39/71/4b/ae1cd004fff64e7ba3c73b384481e9e8/bc1cd004d1f64e7ba3c67b28da81e9e8.65a0ba70.png');
     expect(entity.created).toEqual('2020-03-26T11:14:02+00:00');